Change the execution_time data type to display correctly

+++ src/main/java/com/nameless/lab4/dto/PointResponse.java
@@ -30,5 +30,5 @@ public class PointResponse {
     private String currentTime;
 
     @NotNull
-    private Long executionTime;
+    private Double executionTime;
 }
